More changes for PR: - TFCGuiAreas are now a proper class (Renamed to TFCInventoryGuiHandler as requested). - Chisel recipe wrapper now set output to the debug block if failed to pick block. - Fixed vein registry parsing invalid ore blocks.
